Split state explorer into modules
